refactor(frontend): split playtest visibility runtime
